+++ src/main/java/lotto/LottoMain.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.controller.LottoGame;
+
 public class LottoMain {
 
     public static void main(String[] args) {
