+++ src/main/java/lotto/domain/LottoTickets.java
@@ -1,8 +1,6 @@
-package lotto;
+package lotto.domain;
 
 import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class LottoTickets {
     private List<LottoTicket> lottoTickets;
