+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -1,13 +1,15 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoRank;
+import lotto.domain.LottoTicket;
+import lotto.domain.Number;
+import lotto.domain.WinnerNumber;
 import org.assertj.core.util.Sets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.*;
