+++ src/test/java/lotto/domain/NumberPoolTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.NumberPool;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
