+++ src/test/java/lotto/domain/NumberTest.java
@@ -1,9 +1,8 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.*;
